Migrated to EncryptedSharedPreferences + Extracted Retrofit Interceptor
Token data is sensitive, so after failing to find an encrypted version of DataStore, I opted to simply replace it with EncryptedSharedPreferences.

Previously added token in Hilt Core Module's provideOkHttpClient() fun, but that would only read the token when the app was first built.  Extracted that logic out into AuthorizationInterceptor.kt so that the interceptor can read the token for every call. (Main use case is that the token is empty until the user logs in, at which point they would previously need to reopen the app to get their actual token used.)