Improved TimeNeedle.kt placement logic.
Instead of calculating the index of the item it is supposed to go after and then checking the index on every item in the list we group our list by their startTime in relation to current time. Then, we can have two items{ } blocks in our LazyColumn and simply place the TimeNeedle.kt between them.